Add Error_analysis_evolve.ipynb
This notebook analyzed the error when doing the evolution simulation. When we choose a smaller dt, we can get more accuracy but need more time to run our code. Also, different methods(steppers) have different performances. We want less time complexity with higher accuracy. According to this analysis, we should choose exp-step as the stepper. Also, we may need dt<<T, and dt <1 , in order to have better accuracy.